Update README.md
Q1. Open ClassSizeRKD.dta, this data should look very familiar. It is the exact same as ClassSizeRD.dta that you used in Problem Set #3 except it now contains a new variable named num_classrooms. That variable is equal to the number of classrooms in each grade cohort. You should not need to use the data for parts (a)-(c). Remember to cluster standard errors at the level of coh_id, like you did in problem set #3.

a. Assume that Maimonides Rule is applied as follows. If a grade cohort contains 40 students or less, they are all put together in the same classroom. If a grade cohort contains 41-80 students, they are split into classes of equal size. Based on this, draw a graph of the relationship between cohort size and number of classes in the grade cohort. In this theoretical relationship, what is the discontinuity in number of classes at the cutoff and what is the change in slope at the cutoff?

b. Based on the description in (a), draw a graph of the relationship between cohort size and class size in the grade cohort. In this theoretical relationship, what is the discontinuity in class size at the cutoff and what is the change in slope at the cutoff?

c. When number of classes in a grade cohort increases, schools need to hire more teachers to teach those classes. Assume that school budgets are limited, so when they need to hire more teachers, they pay each teacher less, and the quality of teachers employed for the grade cohort is lower. If this is true, then based on your answers to (a) and (b), does the RDD that you estimated in problem set #4 identify the causal effect of class size on test scores? Why or why not? If not, explain whether this introduces positive or negative bias to your estimate. In other words, does this cause you to underestimate or overestimate the effect of class size on test scores using the RDD in problem set #3? Why?